+++ pytorch_lightning/core/datamodule.py
@@ -19,7 +19,7 @@ from argparse import ArgumentParser, Namespace
 from typing import Any, List, Optional, Tuple, Union
 
 import torch
-from pytorch_lightning.core.hooks import CheckpointHooks, DataHooks
+from pytorch_lightning.core.hooks import DataHooks
 from pytorch_lightning.utilities import parsing, rank_zero_only
 from torch.utils.data import DataLoader
 
@@ -92,7 +92,7 @@ def track_data_hook_calls(fn):
     return wrapped_fn
 
 
-class LightningDataModule(DataHooks, CheckpointHooks, metaclass=_DataModuleWrapper):
+class LightningDataModule(DataHooks, metaclass=_DataModuleWrapper):
     """
     A DataModule standardizes the training, val, test splits, data preparation and transforms.
     The main advantage is consistent data splits, data preparation and transforms across models.
